Update to 1.1.1
