+++ README.md
@@ -1,0 +1,7 @@
+To update:
+
+```
+python3 flatpak-node-generator.py npm --xdg-layout package-lock.json
+python3 update-library.py
+python3 update-packager.py
+```
