+++ org.turbowarp.TurboWarp.yaml
@@ -23,12 +23,13 @@ modules:
     sources:
       - type: git
         url: https://github.com/TurboWarp/desktop.git
-        commit: db0e0b1d18a3b4d49ec04c45366e008f4d6bcbd0
+        commit: 957097fa6bb2197db19f90cc694178562dbf8f0c
       - type: script
         dest-filename: run.sh
         commands:
           - exec zypak-wrapper /app/turbowarp/turbowarp-desktop "$@"
-      - asset-sources.json
+      - library-sources.json
+      - packager-sources.json
       - generated-sources.json
     build-options:
       env:
@@ -42,6 +43,11 @@ modules:
         . ./flatpak-node/electron-builder-arch-args.sh
         npm run dist -- $ELECTRON_BUILDER_ARCH_ARGS --linux --dir
       - cp -a dist/linux*unpacked /app/turbowarp
+      - brotli uncompressed/packager.html
+      - brotli uncompressed/library-files/*
+      - mv uncompressed/packager.html.br static
+      - mkdir -p static/library-files
+      - mv uncompressed/library-files/*.br static/library-files
       - install -Dm755 run.sh /app/bin/turbowarp-desktop
       - install -Dm644 static/icon.png /app/share/icons/hicolor/512x512/apps/org.turbowarp.TurboWarp.png
       - install -Dm644 linux-files/org.turbowarp.TurboWarp.desktop /app/share/applications/org.turbowarp.TurboWarp.desktop
