+++ packager-sources.json
@@ -1,0 +1,9 @@
+[
+    {
+        "type": "file",
+        "url": "https://github.com/TurboWarp/packager/releases/download/v0.6.0/turbowarp-packager-standalone-0.6.0.html",
+        "sha256": "56f0680e3a604a7546ab85d95cc8106b26f4f5c3f69b2319afc155c2caba66e8",
+        "dest": "static",
+        "dest-filename": "packager.html"
+    }
+]
